1. lista doble
constructor de lista doble
