--- storage/team02/Interfaz/Tablas.py
@@ -1,8 +1,0 @@
-
-
-class nodo :
-    def __init__(self,nombreDeLaTabla,numeroDeColumnasDeLaTabla) :
-        self.nombre = nombreDeLaTabla
-        self.columnas = numeroDeColumnasDeLaTabla
-        self.siguiente = None
-        self.anterior = None
